requirements.txt
python-dotenv
pymongo
pymongo[srv]
python-telegram-bot